App/requirements.txt
flask
flask-sqlalchemy
tensorflow
numpy
pillow
gunicorn
werkzeug
